# src/main/java/com/example/store/entity/Order.java
package com.example.store.entity;

import jakarta.persistence.*;

import lombok.Data;

@Entity
@Data
@Table(name = "\"order\"")
public class Order {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private String description;

    @ManyToOne(fetch = FetchType.LAZY)
    private Customer customer;
}